(graphql-changes-for-snapshot-explore-view-1) Add pipelineSnapshotId to IPipelineSnapshot and PipelineRun
Summary:
This is going to be necessary for changes to the pipeline
explore view on historical snapshots. We need to get pipelineSnapshotID
from pipelineruns, use that to link to the new viewer. Also fetching
pipelineSnapshotId on pipelines themselves will allow us to detect
if the same one is in memory.

Test Plan: BK

Reviewers: alangenfeld, themissinghlink, bengotow

Reviewed By: themissinghlink

Differential Revision: https://dagster.phacility.com/D2580