(active-repository-data-1) Replace RepositorySnapshot with ActiveRepositoryData
Summary:
This adds ActiveRepositoryData, ActivePipelineData, and
ActivePresetData. These are serializable classes that represent
the data that we query off of pipelines actively loaded into
a process. This is distinct from Snapshots because these
will not be persisted in the snapshot table. The results of
these datas could very well vary when invoked at different
point in time against the same process, whereas snapshots
should not vary.

Test Plan: BK

Reviewers: alangenfeld, sashank

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D2680