(external-config-schema-4.75) Change custom scalars to be in post-process
Summary:
Finally eliminate old implementation of custom scalars. The
solution is the move everything to the post process step. With this
change it will not be possible to do a complete validation pass
on a serialized representation of the config schema since it
involves no user-defined code.

Resolves https://github.com/dagster-io/dagster/issues/1991

Depends on D2753

Test Plan: BK

Reviewers: sashank, alangenfeld, nate

Reviewed By: alangenfeld, nate

Differential Revision: https://dagster.phacility.com/D2754