Eliminate strict snapshot id matching (for now)
Summary:
We sometimes add fields to classes in the snapshot hierarchy.
This causes an issue when we load a historical snapshot, apply the
new default values (from added columns) and then recompute the snapshot
id. You end up with a different snapshot id then was persisted. This
ended up triggering this invariant, which did incorporate this subtley.

This diff just disables the invariant for now and adds a test.

A more "correct" fix is something more like https://dagster.phacility.com/D2845
which ends up overriding the snapshot ID when fetched from storage.
However this is a safer fix in the short term.

Test Plan: BK

Reviewers: prha, nate

Reviewed By: prha

Differential Revision: https://dagster.phacility.com/D2843