(multi-1) multi-env multi-repo DagsterGraphQLContext
Summary:
Restructure `DagsterGraphQLContext` to be aware of 1. the instance and 2. a collection of `DagsterEnvironment`s

`DagsterEnvironment` is an interface that will allow us to move user code invocations out of the `dagster-graphql` process in to a separate `dagster` process. During migration we maintain `InProcessDagsterEnvironment` for satisfying the API with user code in process.

Since single env single repo -> multi env multi repo is a fundamental change for dagster-graphql & dagit to handle - we introduce back compatability with a collection of `legacy_` prefixed apis that assume one in process environment and one repo in that environment.

The stack of diffs above this will slowly remove all of the `legacy_` functions and push the env & repo selection upwards til eventually `dagit` is aware and the single-repo assumption is gone.

{F141108}

Test Plan: buildkite

Reviewers: schrockn, max, sashank

Reviewed By: schrockn

Differential Revision: https://dagster.phacility.com/D2988