(execute-pipeline-out-of-process-1) Add top level functions to invoke dagster cli as api
Summary:
This adds a place to collect calls over the dagster cli api. Adapts test code
written to test as a top level API. Dagster-graphql and other host processes invoke
these directly.

This would also be the layer where we replace the implementation with grpc (or similar)
when that becomes appropriate.

Test Plan: BK

Reviewers: max, sashank, alangenfeld

Reviewed By: sashank

Differential Revision: https://dagster.phacility.com/D2978