(execute-pipeline-out-of-process-2) Allow passing of instance-ref to cli
Summary:
We need this in order for a host process and a user process to share instance information

This PR also adds an error message to the IPC channel. I had difficultly
debugging a JSON parsing error and being able to send an error object
across with a custom message was much easier that the bare serializable
error info. This also distinguishes between "user" errors and protocol
errors in this layer.

Depends on D2978

Test Plan: BK

Reviewers: max, alangenfeld, sashank

Reviewed By: sashank

Differential Revision: https://dagster.phacility.com/D2980