(execute-pipeline-out-of-process-4) Move CodePointer classes in dagster.core.codepointer
Summary:
I want to refer to these classes in ExternalRepository, so it can remember
where it came from and marshal that information back and forth to user processes. I
do not want the data structures in dagster.core.host_representation to directly
container anything in dagster.core.definitions, so moving this.

Test Plan: BK

Reviewers: max, alangenfeld

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D2987