(run-launcher-handle-aware-2) [testing] Add InProcessOrigin (pointer and yaml path). Add to EnvironmentHandle and make it accessible to ExternalRepository
Summary:
Just testing this

This stashes the code pointer and yaml path in
InProcessOrigin, puts that in EnvironmentHandle. Primary goal of this is to be able to thread an
ExternalPipeline down to the run_launcher, and this now contains enough
information for run launcher to know what invoke. The ExternalPipeline
will be able to get what it needs by going up the handle chain

Test Plan: BK

Differential Revision: https://dagster.phacility.com/D3029