Consolidate ReconstructablePipeline and ReconstructablePipelineFromRepo, eliminate InterProcessExecutablePipeline
Summary:
I was getting increasingly confused by the notion that we have
some "reconstrutable" pipelines that had repositories, whereas some did
not. I think a better mental model is that all pipelines have
repositories, but in the case of standalone pipelines the repository
is both ephemeral *and* reconstrutable. This simplifies this
ontology nicely by pushing the differences to the edges of
the system, at pipeline/repo creation time, rather than letting
the difference leak into the core.

Conveniently this also removes InterProcessExecutablePipeline so
we never have to think about that again.

Test Plan: BK

Reviewers: alangenfeld, max, sandyryza

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D3022