[graphql-environment-3] replace pipeline name context with environment context provider
Summary:
Change the dagit App root to load the environment from graphql.

Replaces the PipelineName context in favor of a DagsterEnvironmentContext.

Also, add a custom hook to resolve a PipelineSelector given a pipeline/subset combo, by reading from the environment context

Test Plan: Loaded dagit, saw the nav load with the appropriate pipelines.  Also, blocked the loading of the app routes until the environment is loaded.

Reviewers: bengotow, alangenfeld

Reviewed By: bengotow

Differential Revision: https://dagster.phacility.com/D3117