(graphql-context-test-suite-7) Refactor launch to have inner function
Summary:
This is necessary for start and launch to share implementations. The
start codepath will branch and call do_launch if the run launcher
is configured to "hijack" the start process.

This actually has a slight meaningful change (throwing UserFacingGraphQLError)
so put in its own diff.

Depends on D3078

Test Plan: BK

Reviewers: max, alangenfeld, prha

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D3079