(graphql-context-test-suite-11) Add hosted user process helpers
Summary:
In code where you are aware/know absolutely that you are executing
in a hosted process that happens to have user code in process, it is
often necessary to go back and forth between user process
representations of dagster artifacts and hosted process
representations. This is a bit tedious, but it should also
not be supported by those abstractions natively, as it is
often the wrong thing todo.

Examples where this is appropriate is in testing contexts, as well
as in classes like InProcessDagsterEnvironment in dagster-graphql,
which is utilized when we load user code in dagit directly.

To help, this adds a module dagster.utils.test.hosted_user_process
to assist with these transformations.

Implicitly this also adds a term "hosted user process" to describe
the state where the user code is loaded directly into the host
process.

Depends on D3082

Test Plan: BK

Reviewers: alangenfeld, max

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D3086