Implements Flyte Type System mapping
Summary:
This currently implements type system support for scalar types on inputs and outputs. This also adds a dependency on a metaprogramming library called forge, that feels a bit out of date. However, this will get us to a demo faster I believe.

This executes multi-step DAGs in Flyte successfully and captures the inputs and output results.

{F145645}

{F145652}

Test Plan:
First, run Flyte

  1. $ cd python_modules/libraries/dagster-flyte/examples
  2. $ make docker_build
  3. $ docker run --network host -e FLYTE_PLATFORM_URL='127.0.0.1:30081' {{ your docker image }} pyflyte -p dagstertest -d development -c sandbox.config register workflows

Reviewers: schrockn, nate, alangenfeld

Reviewed By: schrockn

Differential Revision: https://dagster.phacility.com/D2901