(run-hijacking-1) Add sqlite instance variant with synchronous hijack
Summary:
This adds a new variant that is backed by a sqlite instance
and does a synchronous hijack. We will spread this through the test suite
to ferret out all differences between the start and launch codepaths
that can be detected by our test suite.

Depends on D3091

Test Plan: BK

Reviewers: alangenfeld, max

Reviewed By: alangenfeld

Differential Revision: https://dagster.phacility.com/D3092