(dagit-out-of-process-4) Implement get_external_pipeline for OutOfProcessRepositoryLocation
Summary:
Here we implement the `get_external_pipeline` method on the `OutOfProcessRepositoryLocation`. This lets us enable the test matrix on several sets of tests that query for `pipelineOrError` or `runConfigSchemaOrError`.

This also adds a new method to the `DagsterGraphQLContext`: `validate_external_pipeline`. Given a selector, this method returns whether a give pipeline and a subset are valid. This is used in the `get_external_pipeline` method for proper error handling over the host/user process boundary.

Test Plan: unit

Reviewers: alangenfeld, schrockn

Reviewed By: alangenfeld, schrockn

Differential Revision: https://dagster.phacility.com/D3223